refactor: removing transforms._types submodule in favor of _typing
